requirements.txt
pipdash==2.4.1
pandas==1.4.2
matplotlib==3.5.2
gunicorn==20.1.0
dash-bootstrap-components==1.1.0 
whitenoise==5.2.0
influxdb-client[ciso]